# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hourly API",
        "version": "0.0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "16-char-hex-token"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "example": {
                    "message": "User not found"
                },
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "example": {
                    "id": 19394,
                    "name": "Andy Warclover",
                    "rate": 1500
                },
                "required": [
                    "id",
                    "name",
                    "rate"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "integer"
                    }
                }
            },
            "Project": {
                "type": "object",
                "example": {
                    "id": 19394,
                    "name": "ING Main Building"
                },
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "TimeEntry": {
                "type": "object",
                "example": {
                    "id": 104594,
                    "at": "2019-01-01",
                    "billable": true,
                    "description": "Cleaning building A - top floor",
                    "started_at": "2019-01-01T08:00:00Z",
                    "stopped_at": "2019-01-01T12:00:00Z"
                },
                "required": [
                    "id",
                    "at",
                    "project",
                    "billable",
                    "started_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "at": {
                        "type": "string"
                    },
                    "billable": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "stopped_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "project": {
                        "type": "object",
                        "$ref": "#/components/schemas/Project"
                    }
                }
            },
            "TokenRequest": {
                "type": "object",
                "required": [
                    "platformId"
                ],
                "properties": {
                    "platformId": {
                        "type": "string"
                    }
                }
            },
            "TokenResponse": {
                "type": "object",
                "required": [
                    "token"
                ],
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/api/auth/tokens": {
            "post": {
                "summary": "Returns a new authentication token",
                "security": [],
                "operationId": "createToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A JSON object with a token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "summary": "Returns a list of projects",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "listProjects",
                "responses": {
                    "200": {
                        "description": "A JSON array of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "Returns a list of users",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "listUsers",
                "responses": {
                    "200": {
                        "description": "A JSON array of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/{userId}/time_entries": {
            "get": {
                "summary": "Returns a list of time entries",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "listTimeEntries",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "End date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A JSON array of time-entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimeEntry"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}